Novos métodos funcionais

+++ src/Servico.java
@@ -25,7 +25,7 @@ public class Servico {
         return valor;
     }
 
-    public void setValor(int valor) {
+    public void setValor(double valor) {
         this.valor = valor;
     }
